Implement and use filter to show the first N paragraphs: intro.
This filter replaces the use of the summarize filter. With the latter
it is harder to control what is shown in the entry listings; sometimes
headings also appear in the summary if the first paragraph is short
enough. What I want is to shown the first paragraph of an article. The
new filter does this.