Remove icon font and replace the RSS icon with an SVG.
The only icon left at this point was the RSS icon to draw attention to
the Atom feed. By replacing this with an SVG, we don't have to include
an icon font.

Although this may be a bit less efficient for users that visit more
than one article (since the SVG is included on all pages where it is
used while the font is downloaded once and cached by the browser),
most visitors only read one article anyway.